% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_measurement_index.R
\name{calculate_measurements_index}
\alias{calculate_measurements_index}
\title{Calcula los indices de calidad del aire de un set de mediciones.}
\usage{
calculate_measurements_index(
  measurements,
  hours = 1,
  weighted = FALSE,
  relevant_gap = 3,
  min_relevant_records = 2
)
}
\arguments{
\item{measurements}{Un vector numerico que representa las mediciones para los calculos,
debe de tener al menos un valor. Los valores deben estar ordenados con los mas antiguos primero, mas recientes al
final. Se debe considerar las mediciones deben ser continuas, si existen falta de datos, se debe previamente
rellenar con `NAs`.}

\item{hours}{Cantidad de horas que se considerar치n para el calculo. Por defecto 1.}

\item{weighted}{Indica si el calculo es para concentraci칩n promedio movil ponderado. FALSE por defecto.}

\item{relevant_gap}{Indica cuantos de los registros mas recientes deben considerarse como obligatorios para la
concentracion promedio movil ponderado (weighted = TRUE). No debe ser menor a 0 en caso de usarse, y no debe ser
mayor al valor de hours.}

\item{min_relevant_records}{Indica cuantas de las registros mas recientes (indicados por relevant_gap) son
obligatorios para que se puede hacer el calculo para concentraci칩n promedio m칩vil ponderado.}
}
\value{
Vector numerico que representa el Indice de calidad del aire.
}
\description{
Calcula los indices de calidad del aire de un set de mediciones.
}
\examples{
# Indice horario (1 hora)
index <- calculate_measurements_index(c(1,2,3))

# Indice 8 horas
index <- calculate_measurements_index(c(1,2,3,4,5,6,7,8), hours=8)

# Indice 12 horas ponderado
index <- calculate_measurements_index(
 c(1,2,3,4,5,6,7,8,9,10,11,12),
 hours=12,
 weighted=TRUE,
 relevant_gap=3,
 min_relevant_records=2
)
}
