% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_report_to_data_frame.R
\name{convert_report_to_data_frame}
\alias{convert_report_to_data_frame}
\title{Convert a report to a data frame}
\usage{
convert_report_to_data_frame(report)
}
\arguments{
\item{report}{Report either daily or hourly to format}
}
\description{
Convert a report to a data frame
}
