% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_air_quality_index_data.R
\name{extract_air_quality_index_data}
\alias{extract_air_quality_index_data}
\title{Extraer los datos historicos de los indices de calidad del aire horarios almacenados en una base de datos.}
\usage{
extract_air_quality_index_data(
  connection,
  start_time,
  end_time,
  filter = TRUE,
  sufix_column_names = ""
)
}
\arguments{
\item{connection}{Conexion a la base de datos}

\item{start_time}{Fecha y hora de inicio}

\item{end_time}{Fecha y hora de fin}

\item{filter}{Indica si se aplica el filtro interno que oculta los modulos desactivados internamente.}

\item{sufix_column_names}{Le da un sufijo a las columnas de los datos.}
}
\value{
Devuelve un dataframe con los datos extraidos.
}
\description{
Extraer los datos historicos de los indices de calidad del aire horarios almacenados en una base de datos.
}
