% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_measurements_data.R
\name{extract_measurements_data}
\alias{extract_measurements_data}
\title{Extraer las mediciones en crudo o prevalidadas almacenadas en la base de datos.}
\usage{
extract_measurements_data(
  connection,
  start_time,
  end_time,
  prevalidated = TRUE
)
}
\arguments{
\item{connection}{Conexion a la base de datos}

\item{start_time}{Fecha y hora de inicio}

\item{end_time}{Fecha y hora de fin}

\item{prevalidated}{Indica si los datos son prevalidados o son crudos.}
}
\value{
Devuelve un dataframe con los datos extraidos.
}
\description{
Extraer las mediciones en crudo o prevalidadas almacenadas en la base de datos.
}
