% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hourly_air_quality_index_report.R
\name{generate_hourly_air_quality_index_report}
\alias{generate_hourly_air_quality_index_report}
\title{Genera un reporte horario con los indices de calidad del aire
correspondientes a la hora y fecha indicada.}
\usage{
generate_hourly_air_quality_index_report(
  date_time,
  measurements_data,
  control,
  limits,
  intervals,
  locations,
  parameters,
  categories,
  indexes
)
}
\arguments{
\item{date_time}{Indica un valor de tipo hora y fecha que indica un momento concreto para generar el reporte de
calidad del aire.}

\item{measurements_data}{Es un dataframe con los datos de las estaciones con los distintos parametros que son usados
para calcular los indices de calidad del aire.}

\item{control}{Es un dataframe que indica las estaciones e indices que se van a calcular, así como notar cuáles
modulos estan en mantenimiento o desactivados.}

\item{limits}{Dataframe que indica los valores minimos y maximos de los contaminantes que se usaran para los calculos.}

\item{intervals}{Dataframe que indica los rangos de los indices de calidad del aire y a qué nivel de riesgo o
categoria de calidad del aire corresponden.}

\item{locations}{Tabla que registra las ubicaciones o estaciones en donde se extraen las mediciones.}

\item{parameters}{Indica los parametros o contaminantes e informacion sobre su comportamiento para los calculos.}

\item{categories}{Registro de las categorias o niveles de riesgo de calidad del aire.}

\item{indexes}{Indica cómo o qué consideran los indices para su calculo, asi como a que parametro corresponden y
que ajustes de harán para calcular el indice de calidad del aire.}
}
\value{
Devuelve una lista con la hora y fecha del reporte, las ubicaciones consideradas, los parametros o contaminantes,
las categorias de calidad del aire, los indices y que consideran para el calculo, y los resultados, que contienen
todos los resultados de cada indice calculado.
}
\description{
Genera un reporte horario con los indices de calidad del aire
correspondientes a la hora y fecha indicada.
}
